Use localstorage for login
